refactor(pokeapi): move pokemon type
