+++ pokedex/src/pokeapi/types_locations.go
@@ -63,65 +63,3 @@ type Location struct {
 		} `json:"version_details"`
 	} `json:"pokemon_encounters"`
 }
-
-// Pokemon
-type Pokemon struct {
-	ID             int    `json:"id"`
-	Name           string `json:"name"`
-	BaseExperience int    `json:"base_experience"`
-	Height         int    `json:"height"`
-	IsDefault      bool   `json:"is_default"`
-	Order          int    `json:"order"`
-	Weight         int    `json:"weight"`
-	Abilities      []struct {
-		IsHidden bool `json:"is_hidden"`
-		Slot     int  `json:"slot"`
-		Ability  struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"ability"`
-	} `json:"abilities"`
-	Forms []struct {
-		Name string `json:"name"`
-		URL  string `json:"url"`
-	} `json:"forms"`
-	GameIndices []struct {
-		GameIndex int `json:"game_index"`
-		Version   struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"version"`
-	} `json:"game_indices"`
-	HeldItems []struct {
-		Item struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"item"`
-		VersionDetails []struct {
-			Rarity  int `json:"rarity"`
-			Version struct {
-				Name string `json:"name"`
-				URL  string `json:"url"`
-			} `json:"version"`
-		} `json:"version_details"`
-	} `json:"held_items"`
-	LocationAreaEncounters string `json:"location_area_encounters"`
-	Moves                  []struct {
-		Move struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"move"`
-		VersionGroupDetails []struct {
-			LevelLearnedAt int `json:"level_learned_at"`
-			VersionGroup   struct {
-				Name string `json:"name"`
-				URL  string `json:"url"`
-			} `json:"version_group"`
-			MoveLearnMethod struct {
-				Name string `json:"name"`
-				URL  string `json:"url"`
-			} `json:"move_learn_method"`
-		} `json:"version_group_details"`
-	} `json:"moves"`
-}
-
